# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GTrader API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all trading accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new trading account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a trading account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update a trading account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete a trading account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /accounts/{id}/stats:
    get:
      operationId: getAccountStats
      tags: [accounts]
      summary: Get statistics for a specific account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brokerName:
          type: ["string", "null"]
        accountType:
          type: string
          description: "forex, crypto, prop_firm, demo, cent"
        initialBalance:
          type: number
        currentBalance:
          type: number
        growthTarget:
          type: ["number", "null"]
        dailyLossLimit:
          type: ["number", "null"]
        maxDrawdown:
          type: ["number", "null"]
        maxTradesPerDay:
          type: ["integer", "null"]
        status:
          type: string
          description: "active, passed, failed, blown, archived"
        currency:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - accountType
        - initialBalance
        - currentBalance
        - status
        - currency
        - createdAt
        - updatedAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        brokerName:
          type: ["string", "null"]
        accountType:
          type: string
        initialBalance:
          type: number
        currentBalance:
          type: number
        growthTarget:
          type: ["number", "null"]
        dailyLossLimit:
          type: ["number", "null"]
        maxDrawdown:
          type: ["number", "null"]
        maxTradesPerDay:
          type: ["integer", "null"]
        status:
          type: string
        currency:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - accountType
        - initialBalance
        - currentBalance
        - status
        - currency
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        brokerName:
          type: ["string", "null"]
        accountType:
          type: ["string", "null"]
        currentBalance:
          type: ["number", "null"]
        growthTarget:
          type: ["number", "null"]
        dailyLossLimit:
          type: ["number", "null"]
        maxDrawdown:
          type: ["number", "null"]
        maxTradesPerDay:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AccountStats:
      type: object
      properties:
        accountId:
          type: integer
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        breakevens:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        averageWin:
          type: number
        averageLoss:
          type: number
        profitFactor:
          type: number
        averageRR:
          type: number
        maxDrawdown:
          type: number
        disciplineScore:
          type: number
        consistencyScore:
          type: number
        bestSession:
          type: ["string", "null"]
        bestSetup:
          type: ["string", "null"]
        bestTimeframe:
          type: ["string", "null"]
      required:
        - accountId
        - totalTrades
        - wins
        - losses
        - breakevens
        - winRate
        - totalPnl
        - averageWin
        - averageLoss
        - profitFactor
        - averageRR
        - maxDrawdown
        - disciplineScore
        - consistencyScore
...